enable single client auto iperf (redoing failed trials)

+++ helper_scripts/auto_iperf.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
 
-if [ "$#" -ne 3 ]; then
-    echo "Usage: $0 <num_nodes> <bandwidth> <tdd configuration>"
+if [ "$#" -lt 3 ] || [ "$#" -gt 4 ]; then
+    echo "Usage: $0 <num_nodes|specific_node> <bandwidth> <tdd configuration> [single_client_id]"
+    echo "Examples:"
+    echo "  $0 6 40 7-2              # Run for 6 clients"
+    echo "  $0 6 40 7-2 3            # Run only for client 3"
     exit 1
 fi
 
 NUM_CLIENTS=$1
 BW=$2
 TDD=$3
+SINGLE_CLIENT=${4:-0}  # Optional 4th argument, default to 0 (run all)
 
 DEVICE_NAME_PREFIX="commnetpi0"
 IP_PREFIX="129.105.6."
@@ -18,19 +22,56 @@ TIME=30 #iperf test length (seconds)
 mkdir -p iperf
 mkdir -p "iperf/${BW}_${TDD}/"
 
-# echo "disconnect all devices from 5G"
-# for ((CID=1;CID<=NUM_CLIENTS;CID++)); do
-#     LOGIN=$DEVICE_NAME_PREFIX$CID@$IP_PREFIX${IP_SUFFIXES[$CID-1]}
-#     ssh "$LOGIN" sudo nmcli connection down OAI
-#     ssh "$LOGIN" sudo nmcli r wwan off
-# done
+# Check if CSV files exist and prompt user
+if [ -f "iperf/${BW}_${TDD}/UL.csv" ] || [ -f "iperf/${BW}_${TDD}/DL.csv" ]; then
+    echo "CSV files already exist for ${BW}_${TDD}"
+    read -p "Do you want to (o)verwrite or (a)ppend? [o/a]: " choice
+    case $choice in
+        [Aa]* )
+            APPEND_MODE=true
+            echo "Data will be appended to existing CSV files"
+            ;;
+        [Oo]* )
+            APPEND_MODE=false
+            echo "Existing CSV files will be overwritten"
+            echo "start,ns,device,interval(s),transfer(MB),bandwidth(Mbps),burst_lat_avg(ms),burst_lat_min(ms),burst_lat_max(ms),burst_lat_stdev(ms),cnt,size,inP,NetPwr,Reads=Dist,end,ns" > "iperf/${BW}_${TDD}/UL.csv"
+            echo "start,ns,device,interval(s),transfer(MB),bandwidth(Mbps),burst_lat_avg(ms),burst_lat_min(ms),burst_lat_max(ms),burst_lat_stdev(ms),cnt,size,inP,NetPwr,Reads=Dist,end,ns" > "iperf/${BW}_${TDD}/DL.csv"
+            ;;
+        * )
+            echo "Invalid choice. Exiting."
+            exit 1
+            ;;
+    esac
+else
+    APPEND_MODE=false
+    echo "start,ns,device,interval(s),transfer(MB),bandwidth(Mbps),burst_lat_avg(ms),burst_lat_min(ms),burst_lat_max(ms),burst_lat_stdev(ms),cnt,size,inP,NetPwr,Reads=Dist,end,ns" > "iperf/${BW}_${TDD}/UL.csv"
+    echo "start,ns,device,interval(s),transfer(MB),bandwidth(Mbps),burst_lat_avg(ms),burst_lat_min(ms),burst_lat_max(ms),burst_lat_stdev(ms),cnt,size,inP,NetPwr,Reads=Dist,end,ns" > "iperf/${BW}_${TDD}/DL.csv"
+fi
 
-echo "start,ns,device,interval(s),transfer(MB),bandwidth(Mbps),burst_lat_avg(ms),burst_lat_min(ms),burst_lat_max(ms),burst_lat_stdev(ms),cnt,size,inP,NetPwr,Reads=Dist,end,ns" > "iperf/${BW}_${TDD}/UL.csv"
-echo "start,ns,device,interval(s),transfer(MB),bandwidth(Mbps),burst_lat_avg(ms),burst_lat_min(ms),burst_lat_max(ms),burst_lat_stdev(ms),cnt,size,inP,NetPwr,Reads=Dist,end,ns" > "iperf/${BW}_${TDD}/DL.csv"
+# Determine which clients to test
+if [ "$SINGLE_CLIENT" -gt 0 ]; then
+    echo "Running test for single client: pi0${SINGLE_CLIENT}"
+    START_CLIENT=$SINGLE_CLIENT
+    END_CLIENT=$SINGLE_CLIENT
+else
+    echo "Running test for clients 1 through ${NUM_CLIENTS}"
+    START_CLIENT=1
+    END_CLIENT=$NUM_CLIENTS
+fi
+
+# # Disconnect all devices from 5G (optional, only if running all clients)
+# if [ "$SINGLE_CLIENT" -eq 0 ]; then
+#     echo "Disconnecting all devices from 5G"
+#     for ((CID=1;CID<=NUM_CLIENTS;CID++)); do
+#         LOGIN=$DEVICE_NAME_PREFIX$CID@$IP_PREFIX${IP_SUFFIXES[$CID-1]}
+#         ssh "$LOGIN" sudo nmcli connection down OAI
+#         ssh "$LOGIN" sudo nmcli r wwan off
+#     done
+# fi
 
 # i could also automate oaibox setup and teardown through running oaibox offline...
 
-for ((CID=1;CID<=NUM_CLIENTS;CID++)); do
+for ((CID=START_CLIENT;CID<=END_CLIENT;CID++)); do
     LOGIN=$DEVICE_NAME_PREFIX$CID@$IP_PREFIX${IP_SUFFIXES[$CID-1]}
     echo
     echo $LOGIN
@@ -44,10 +85,10 @@ for ((CID=1;CID<=NUM_CLIENTS;CID++)); do
     # downlink test
     ssh "$LOGIN" "iperf -s -i 1 -fm -B ${wwan} -t ${TIME}" > "iperf/${BW}_${TDD}/pi0${CID}_DL.txt" &
     sleep 2
-    start=$(gdate -u -Ins)
+    start=$(date -u -Ins)
     echo start ${start}
     ssh user@10.105.46.208 docker exec oai-ext-dn iperf -t ${TIME} -i 1 -fm -B 172.31.0.135 -p 5001 -c ${wwan} --trip-times #--txdelay-time 1
-    end=$(gdate -u -Ins)
+    end=$(date -u -Ins)
     # scp "$LOGIN":"${BW}_${TDD}_pi0${CID}_DL.txt" "iperf/${BW}_${TDD}/pi0${CID}_DL.txt"
     # ssh  > "iperf/${BW}_${TDD}/pi0${CID}_DL.txt"
     echo end ${end}
@@ -73,10 +114,10 @@ for ((CID=1;CID<=NUM_CLIENTS;CID++)); do
     # uplink test
     ssh user@10.105.46.208 docker exec oai-ext-dn iperf -s -i 1 -B 172.31.0.135 -p 5001 -t ${TIME} -fm > "iperf/${BW}_${TDD}/pi0${CID}_UL.txt" &
     sleep 2
-    start=$(gdate -u -Ins)
+    start=$(date -u -Ins)
     echo start ${start}
-    ssh "$LOGIN" iperf -t ${TIME} -i 1 -fm -c 172.31.0.135 -B "$wwan" --trip-times --txdelay-time 3
-    end=$(gdate -u -Ins)
+    ssh "$LOGIN" iperf -t ${TIME} -i 1 -fm -c 172.31.0.135 -B "$wwan" --trip-times #--txdelay-time 3
+    end=$(date -u -Ins)
     echo end ${end}
     
     # Extract all values from the final summary row
@@ -102,5 +143,4 @@ for ((CID=1;CID<=NUM_CLIENTS;CID++)); do
     sleep 1
 done
 
-# allow 1 specific client in case of missing data
-# put txt files in folder
+# need to parse phys layer metrics for collected data
